refactored more graph code into interfaces
