+++ hypergraphx/core/I_undirected_hypergraph.py
@@ -119,209 +119,45 @@ class IUndirectedHypergraph(IHypergraph):
         """Set the adjacency dictionary."""
         self._adj = adj_dict
 
-    # =============================================================================
-    # Abstract Methods (Must be implemented by subclasses)
-    # =============================================================================
-    
-    @abstractmethod
-    def add_edge(self, edge, *args, **kwargs) -> None:
+    def get_incident_edges(self, node, order: int = None, size: int = None):
         """
-        Add an edge to the hypergraph.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: add_edge(edge, weight=None, metadata=None)
-        - MultiplexHypergraph: add_edge(edge, layer, weight=None, metadata=None)
-        - TemporalHypergraph: add_edge(edge, time, weight=None, metadata=None)
-        """
-        pass
-
-    @abstractmethod
-    def add_edges(self, edge_list, *args, **kwargs) -> None:
-        """
-        Add multiple edges to the hypergraph.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: add_edges(edge_list, weights=None, metadata=None)
-        - MultiplexHypergraph: add_edges(edge_list, edge_layer, weights=None, metadata=None)
-        - TemporalHypergraph: add_edges(edge_list, time_list, weights=None, metadata=None)
-        """
-        pass
-
-    @abstractmethod
-    def remove_edge(self, edge, *args, **kwargs) -> None:
-        """
-        Remove an edge from the hypergraph.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: remove_edge(edge)
-        - MultiplexHypergraph: remove_edge(edge) where edge is ((nodes...), layer)
-        - TemporalHypergraph: remove_edge(edge, time)
-        """
-        pass
-
-    @abstractmethod
-    def get_edges(self, *args, **kwargs):
-        """
-        Get edges from the hypergraph.
-        
-        Note: Parameters vary by implementation due to different filtering capabilities.
-        """
-        pass
-
-    @abstractmethod
-    def check_edge(self, edge, *args, **kwargs) -> bool:
-        """
-        Check if an edge exists in the hypergraph.
-        
-        Note: Signature varies by implementation.
-        """
-        pass
-
-    @abstractmethod
-    def get_weight(self, edge, *args, **kwargs):
-        """
-        Get the weight of an edge.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: get_weight(edge)
-        - MultiplexHypergraph: get_weight(edge, layer)
-        - TemporalHypergraph: get_weight(edge, time)
-        """
-        pass
-
-    @abstractmethod
-    def set_weight(self, edge, *args, **kwargs) -> None:
-        """
-        Set the weight of an edge.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: set_weight(edge, weight)
-        - MultiplexHypergraph: set_weight(edge, layer, weight)
-        - TemporalHypergraph: set_weight(edge, time, weight)
-        """
-        pass
-
-    @abstractmethod
-    def get_weights(self, *args, **kwargs):
-        """
-        Get weights of edges in the hypergraph.
-        
-        Note: Parameters vary by implementation.
-        """
-        pass
-
-    @abstractmethod
-    def get_sizes(self) -> List[int]:
-        """
-        Get the size of each edge in the hypergraph.
-
-        Returns
-        -------
-        list
-            A list of integers representing the size of each edge.
-        """
-        pass
-
-    @abstractmethod
-    def get_orders(self) -> List[int]:
-        """
-        Get the order of each edge in the hypergraph.
-
-        Returns
-        -------
-        list
-            A list of integers representing the order of each edge.
-        """
-        pass
-
-    # =============================================================================
-    # Abstract Metadata Methods (Implementation-specific signatures)
-    # =============================================================================
-    
-    @abstractmethod
-    def get_edge_metadata(self, edge, *args, **kwargs):
-        """
-        Get metadata for a specific edge.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: get_edge_metadata(edge)
-        - MultiplexHypergraph: get_edge_metadata(edge, layer)
-        - TemporalHypergraph: get_edge_metadata(edge, time)
-        """
-        pass
-
-    @abstractmethod
-    def set_edge_metadata(self, edge, *args, **kwargs):
-        """
-        Set metadata for a specific edge.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: set_edge_metadata(edge, metadata)
-        - MultiplexHypergraph: set_edge_metadata(edge, layer, metadata)
-        - TemporalHypergraph: set_edge_metadata(edge, time, metadata)
-        """
-        pass
-
-    @abstractmethod
-    def set_attr_to_edge_metadata(self, edge, *args, field, value):
-        """
-        Set an attribute in edge metadata.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: set_attr_to_edge_metadata(edge, field, value)
-        - MultiplexHypergraph: set_attr_to_edge_metadata(edge, layer, field, value)
-        - TemporalHypergraph: set_attr_to_edge_metadata(edge, time, field, value)
-        """
-        pass
-
-    @abstractmethod
-    def remove_attr_from_edge_metadata(self, edge, *args, field):
-        """
-        Remove an attribute from edge metadata.
-        
-        Note: Signature varies by implementation:
-        - Hypergraph: remove_attr_from_edge_metadata(edge, field)
-        - MultiplexHypergraph: remove_attr_from_edge_metadata(edge, layer, field)
-        - TemporalHypergraph: remove_attr_from_edge_metadata(edge, time, field)
-        """
-        pass
-
-    # =============================================================================
-    # Serialization Support (Abstract - implementation-specific)
-    # =============================================================================
-    
-    @abstractmethod
-    def expose_data_structures(self) -> Dict:
-        """
-        Expose the internal data structures of the hypergraph for serialization.
-
-        Returns
-        -------
-        dict
-            A dictionary containing all internal attributes of the hypergraph.
-        """
-        pass
-
-    @abstractmethod
-    def populate_from_dict(self, data: Dict) -> None:
-        """
-        Populate the attributes of the hypergraph from a dictionary.
+        Get the incident hyperedges of a node in the hypergraph.
 
         Parameters
         ----------
-        data : dict
-            A dictionary containing the attributes to populate the hypergraph.
-        """
-        pass
-
-    @abstractmethod
-    def expose_attributes_for_hashing(self) -> Dict:
-        """
-        Expose relevant attributes for hashing.
+        node : object
+            The node of interest.
+        order : int
+            The order of the hyperedges to consider.
+        size : int
+            The size of the hyperedges to consider.
 
         Returns
         -------
-        dict
-            A dictionary containing key attributes for hashing.
+        list
+            The incident hyperedges of the node.
+
+        Raises
+        ------
+        ValueError
+            If the node is not in the hypergraph.
+
         """
-        pass
+        if node not in self._adj:
+            raise ValueError("Node {} not in hypergraph.".format(node))
+        if order is not None and size is not None:
+            raise ValueError("Order and size cannot be both specified.")
+        if order is None and size is None:
+            return list(
+                [self._reverse_edge_list[edge_id] for edge_id in self._adj[node]]
+            )
+        else:
+            if order is None:
+                order = size - 1
+            return list(
+                [
+                    self._reverse_edge_list[edge_id]
+                    for edge_id in self._adj[node]
+                    if len(self._reverse_edge_list[edge_id]) - 1 == order
+                ]
+            )
