+++ hypergraphx/core/directed_hypergraph.py
@@ -182,54 +182,6 @@ class DirectedHypergraph(IHypergraph):
         """
         return node in self._adj_source and node in self._adj_target
 
-    def get_neighbors(self, node, order: int = None, size: int = None):
-        """
-        Get the neighbors of a node in the hypergraph.
-
-        Parameters
-        ----------
-        node : object
-            The node of interest.
-        order : int
-            The order of the hyperedges to consider.
-        size : int
-            The size of the hyperedges to consider.
-
-        Returns
-        -------
-        set
-            The neighbors of the node.
-
-        Raises
-        ------
-        ValueError
-            If order and size are both specified or neither are specified.
-        """
-        if node not in self._adj_source.keys():
-            raise ValueError("Node {} not in hypergraph.".format(node))
-        if node not in self._adj_target.keys():
-            raise ValueError("Node {} not in hypergraph.".format(node))
-        if order is not None and size is not None:
-            raise ValueError("Order and size cannot be both specified.")
-        if order is None and size is None:
-            neigh = set()
-            edges = self.get_incident_edges(node)
-            for edge in edges:
-                neigh.update(edge)
-            if node in neigh:
-                neigh.remove(node)
-            return neigh
-        else:
-            if order is None:
-                order = size - 1
-            neigh = set()
-            edges = self.get_incident_edges(node, order=order)
-            for edge in edges:
-                neigh.update(edge)
-            if node in neigh:
-                neigh.remove(node)
-            return neigh
-
     def get_incident_edges(self, node, order: int = None, size: int = None):
         """
         Get the incident edges of a node.
@@ -567,204 +519,6 @@ class DirectedHypergraph(IHypergraph):
         for edge in edge_list:
             self.remove_edge(edge)
 
-    def set_edge_list(self, edge_list):
-        self._edge_list = edge_list
-
-    def get_edge_list(self):
-        return self._edge_list
-
-    """def add_empty_edge(self, name, metadata):
-        pass
-        Don't know if needed    
-    """
-
-    def check_edge(self, edge: Tuple[Tuple, Tuple]):
-        """Checks if the specified edge is in the hypergraph.
-
-        Parameters
-        ----------
-        edge : tuple
-            The edge to check.
-
-        Returns
-        -------
-        bool
-            True if the edge is in the hypergraph, False otherwise.
-
-        """
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        return edge in self._edge_list
-
-    # Weight
-    def get_weight(self, edge: Tuple[Tuple, Tuple]):
-        """Returns the weight of the specified directed edge."""
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge in self._edge_list:
-            idx = self._edge_list[edge]
-            return self._weights[idx]
-        else:
-            raise ValueError(f"Edge {edge} not in hypergraph.")
-
-    def get_weights(self, order=None, size=None, up_to=False, asdict=False):
-        """Returns the list of weights of the edges in the hypergraph. If order is specified, it returns the list of weights of the edges of the specified order.
-        If size is specified, it returns the list of weights of the edges of the specified size. If both order and size are specified, it raises a ValueError.
-        If up_to is True, it returns the list of weights of the edges of order smaller or equal to the specified order.
-
-        Parameters
-        ----------
-        order : int, optional
-            Order of the edges to get the weights of.
-
-        size : int, optional
-            Size of the edges to get the weights of.
-
-        up_to : bool, optional
-            If True, it returns the list of weights of the edges of order smaller or equal to the specified order. Default is False.
-
-        Returns
-        -------
-        list
-            List of weights of the edges in the hypergraph.
-
-        Raises
-        ------
-        ValueError
-            If both order and size are specified.
-
-        """
-        w = None
-        if order is not None and size is not None:
-            raise ValueError("Order and size cannot be both specified.")
-        if order is None and size is None:
-            w = {
-                edge: self._weights[self._edge_list[edge]] for edge in self.get_edges()
-            }
-
-        if size is not None:
-            order = size - 1
-
-        if w is None:
-            w = {
-                edge: self._weights[self._edge_list[edge]]
-                for edge in self.get_edges(order=order, up_to=up_to)
-            }
-
-        if asdict:
-            return w
-        else:
-            return list(w.values())
-
-    def set_weight(self, edge: Tuple[Tuple, Tuple], weight: float):
-        """Sets the weight of the specified directed edge."""
-        if not self._weighted and weight != 1:
-            raise ValueError(
-                "If the hypergraph is not weighted, weight can be 1 or None."
-            )
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge in self._edge_list:
-            idx = self._edge_list[edge]
-            self._weights[idx] = weight
-        else:
-            raise ValueError(f"Edge {edge} not in hypergraph.")
-
-    # Info
-    def num_nodes(self):
-        """Returns the number of nodes in the hypergraph."""
-        return len(self.get_nodes())
-
-    def num_edges(self):
-        """Returns the number of directed edges in the hypergraph."""
-        return len(self._edge_list)
-
-    def get_sizes(self):
-        """Returns the list of sizes of the hyperedges in the hypergraph.
-
-        Returns
-        -------
-        list
-            List of sizes of the hyperedges in the hypergraph.
-
-        """
-        return [len(edge[0]) + len(edge[1]) for edge in self._edge_list.keys()]
-
-    def max_size(self):
-        """
-        Returns the maximum size of the hypergraph.
-
-        Returns
-        -------
-        int
-            Maximum size of the hypergraph.
-        """
-        return max(self.get_sizes())
-
-    def max_order(self):
-        """
-        Returns the maximum order of the hypergraph.
-
-        Returns
-        -------
-        int
-            Maximum order of the hypergraph.
-        """
-        return self.max_size() - 1
-
-    def distribution_sizes(self):
-        """
-        Returns the distribution of sizes of the hyperedges in the hypergraph.
-
-        Returns
-        -------
-        collections.Counter
-            Distribution of sizes of the hyperedges in the hypergraph.
-        """
-        from collections import Counter
-
-        return dict(Counter(self.get_sizes()))
-
-    def get_orders(self):
-        """Returns the list of orders of the hyperedges in the hypergraph.
-
-        Returns
-        -------
-        list
-            List of orders of the hyperedges in the hypergraph.
-
-        """
-        return [len(edge[0]) + len(edge[1]) - 1 for edge in self._edge_list.keys()]
-
-    def is_weighted(self):
-        """
-        Check if the hypergraph is weighted.
-
-        Returns
-        -------
-        bool
-            True if the hypergraph is weighted, False otherwise.
-        """
-        return self._weighted
-
-    def is_uniform(self):
-        """
-        Check if the hypergraph is uniform, i.e. all hyperedges have the same size.
-
-        Returns
-        -------
-        bool
-            True if the hypergraph is uniform, False otherwise.
-        """
-        uniform = True
-        sz = None
-        for edge in self._edge_list:
-            edge = set(edge[0]).union(set(edge[1]))
-            if sz is None:
-                sz = len(edge)
-            else:
-                if len(edge) != sz:
-                    uniform = False
-                    break
-        return uniform
-
     # Adj
     def get_adj_dict(self, source_target):
         if source_target == "source":
@@ -785,22 +539,6 @@ class DirectedHypergraph(IHypergraph):
             raise ValueError(
                 "Invalid value for source_target. Must be 'source' or 'target'."
             )
-
-    # Degree
-    def degree(self, node, order=None, size=None):
-        from hypergraphx.measures.degree import degree
-
-        return degree(self, node, order=order, size=size)
-
-    def degree_sequence(self, order=None, size=None):
-        from hypergraphx.measures.degree import degree_sequence
-
-        return degree_sequence(self, order=order, size=size)
-
-    def degree_distribution(self, order=None, size=None):
-        from hypergraphx.measures.degree import degree_distribution
-
-        return degree_distribution(self, order=order, size=size)
 
     # Connected Components
     """def is_connected(self, size=None, order=None):
@@ -895,56 +633,10 @@ class DirectedHypergraph(IHypergraph):
         return directed_line_graph(self, distance, s, weighted)
 
     # Metadata
-    def set_hypergraph_metadata(self, metadata):
-        self._hypergraph_metadata = metadata
-
-    def get_hypergraph_metadata(self):
-        return self._hypergraph_metadata
-
-    def set_node_metadata(self, node, metadata):
-        if node not in self._adj_source:
-            raise ValueError("Node {} not in hypergraph.".format(node))
-        self._node_metadata[node] = metadata
-
-    def get_node_metadata(self, node):
-        if node not in self._adj_source:
-            raise ValueError("Node {} not in hypergraph.".format(node))
-        return self._node_metadata[node]
-
-    def get_all_nodes_metadata(self):
-        return list(self._node_metadata.values())
-
-    def set_edge_metadata(self, edge, metadata):
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        idx = self._edge_list[edge]
-        self._edge_metadata[idx] = metadata
-
-    def get_edge_metadata(self, edge):
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        idx = self._edge_list[edge]
-        return self._edge_metadata[idx]
-
-    def get_all_edges_metadata(self):
-        return self._edge_metadata
-
-    def set_incidence_metadata(self, edge, node, metadata):
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        self._incidences_metadata[(edge, node)] = metadata
-
-    def get_incidence_metadata(self, edge, node):
-        edge = (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        if edge not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        return self._incidences_metadata[(edge, node)]
-
-    def get_all_incidences_metadata(self):
-        return {k: v for k, v in self._incidences_metadata.items()}
+    def _restructure_query_edge(self, edge):
+        # all the methods from IHypergraph that use this
+        #   already get the canonical form of the edge
+        return edge
 
     def set_attr_to_hypergraph_metadata(self, field, value):
         self._hypergraph_metadata[field] = value
