+++ hypergraphx/core/multiplex_hypergraph.py
@@ -1,21 +1,6 @@
 from hypergraphx import Hypergraph
 from hypergraphx.core.I_undirected_hypergraph import IUndirectedHypergraph
 
-
-def _canon_edge(edge):
-    edge = tuple(edge)
-
-    if len(edge) == 2:
-        if isinstance(edge[0], tuple) and isinstance(edge[1], tuple):
-            # Sort the inner tuples and return
-            return (tuple(sorted(edge[0])), tuple(sorted(edge[1])))
-        elif not isinstance(edge[0], tuple) and not isinstance(edge[1], tuple):
-            # Sort the edge itself if it contains IDs (non-tuple elements)
-            return tuple(sorted(edge))
-
-    return tuple(sorted(edge))
-
-
 class MultiplexHypergraph(IUndirectedHypergraph):
     """
     A Multiplex Hypergraph is a hypergraph where hyperedges are organized into multiple layers.
@@ -107,51 +92,21 @@ class MultiplexHypergraph(IUndirectedHypergraph):
                 metadata=edge_metadata,
             )
 
-    def get_adj_dict(self):
-        return self._adj
-
-    def set_adj_dict(self, adj_dict):
-        self._adj = adj_dict
 
     def get_incident_edges(self, node):
         if node not in self._adj:
             raise ValueError("Node {} not in hypergraph.".format(node))
         return [self._reverse_edge_list[e_id] for e_id in self._adj[node]]
 
-    def degree(self, node, order=None, size=None):
-        from hypergraphx.measures.degree import degree
-
-        return degree(self, node, order=order, size=size)
-
-    def degree_sequence(self, order=None, size=None):
-        from hypergraphx.measures.degree import degree_sequence
-
-        return degree_sequence(self, order=order, size=size)
-
-    def get_edge_metadata(self, edge, layer):
-        edge = tuple(sorted(edge))
-        k = (edge, layer)
-        if k not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        return self._edge_metadata[self._edge_list[k]]
-
-    def get_edge_list(self):
-        return self._edge_list
-
-    def set_edge_list(self, edge_list):
-        self._edge_list = edge_list
+    def _restructure_query_edge(self, edge, layer):
+        # modify the edge being queried by making a tuple of it and the layer
+        return (edge, layer)
 
     def get_existing_layers(self):
         return self._existing_layers
 
     def set_existing_layers(self, existing_layers):
         self._existing_layers = existing_layers
-
-    def get_nodes(self, metadata=False):
-        if metadata:
-            return self._node_metadata
-        else:
-            return list(self._node_metadata.keys())
 
     def add_edges(self, edge_list, edge_layer, weights=None, metadata=None):
         """Add a list of hyperedges to the hypergraph. If a hyperedge is already in the hypergraph, its weight is updated.
@@ -280,8 +235,8 @@ class MultiplexHypergraph(IUndirectedHypergraph):
         ValueError
             If the edge is not in the hypergraph.
         """
-        edge = _canon_edge(edge)
-        if edge not in self._edge_list:
+        edge = self._canon_edge(edge)
+        if not self.check_edge(edge):
             raise ValueError(f"Edge {edge} not in hypergraph.")
 
         edge_id = self._edge_list[edge]
@@ -352,26 +307,6 @@ class MultiplexHypergraph(IUndirectedHypergraph):
         else:
             return list(self._edge_list.keys())
 
-    def get_weight(self, edge, layer):
-        edge = _canon_edge(edge)
-        k = (edge, layer)
-        if k not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(k))
-        else:
-            return self._weights[self._edge_list[k]]
-
-    def set_weight(self, edge, layer, weight):
-        if not self._weighted and weight != 1:
-            raise ValueError(
-                "If the hypergraph is not weighted, weight can be 1 or None."
-            )
-
-        k = (_canon_edge(edge), layer)
-        if k not in self._edge_list:
-            raise ValueError("Edge {} not in hypergraph.".format(edge))
-        else:
-            self._weights[self._edge_list[k]] = weight
-
     def set_dataset_metadata(self, metadata):
         self._hypergraph_metadata["multiplex_metadata"] = metadata
 
@@ -385,12 +320,6 @@ class MultiplexHypergraph(IUndirectedHypergraph):
 
     def get_layer_metadata(self, layer_name):
         return self._hypergraph_metadata[layer_name]
-
-    def get_hypergraph_metadata(self):
-        return self._hypergraph_metadata
-
-    def set_hypergraph_metadata(self, metadata):
-        self._hypergraph_metadata = metadata
 
     def aggregated_hypergraph(self):
         h = Hypergraph(
@@ -406,14 +335,6 @@ class MultiplexHypergraph(IUndirectedHypergraph):
                 metadata=self.get_edge_metadata(_edge, layer),
             )
         return h
-
-    def set_attr_to_hypergraph_metadata(self, field, value):
-        self._hypergraph_metadata[field] = value
-
-    def set_attr_to_node_metadata(self, node, field, value):
-        if node not in self._node_metadata:
-            raise ValueError("Node {} not in hypergraph.".format(node))
-        self._node_metadata[node][field] = value
 
     def set_attr_to_edge_metadata(self, edge, layer, field, value):
         edge = _canon_edge(edge)
@@ -431,6 +352,18 @@ class MultiplexHypergraph(IUndirectedHypergraph):
         if edge not in self._edge_metadata:
             raise ValueError("Edge {} not in hypergraph.".format(edge))
         del self._edge_metadata[self._edge_list[(edge, layer)]][field]
+
+    def clear(self):
+        self._adj.clear()
+        self._weights.clear()
+        self._hypergraph_metadata.clear()
+        self._incidences_metadata.clear()
+        self._node_metadata.clear()
+        self._edge_metadata.clear()
+        self._edge_list.clear()
+        self._reverse_edge_list.clear()
+        self._existing_layers.clear()
+
 
     def expose_data_structures(self):
         """
